% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPconf-methods.R
\name{APPconf-methods}
\alias{APPconf-methods}
\alias{show,APPconf-method}
\alias{appconf}
\alias{<-}
\alias{readRDS(system.file("extdata",}
\alias{"data",}
\alias{"pbmc_small",}
\alias{"appconf.rds",}
\alias{package="scRNAseqApp"))}
\alias{$,APPconf-method}
\alias{$<-,APPconf-method}
\alias{[[,APPconf,ANY,ANY-method}
\alias{[[<-,APPconf,ANY,ANY,ANY-method}
\title{The methods for \link{APPconf-class}}
\usage{
\S4method{show}{APPconf}(object)

\S4method{$}{APPconf}(x, name)

\S4method{$}{APPconf}(x, name) <- value

\S4method{[[}{APPconf,ANY,ANY}(x, i, j, ..., exact = TRUE)

\S4method{[[}{APPconf,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{object}{an object of APPconf}

\item{x}{APPconf object.}

\item{name}{A literal character string or a name (possibly backtick quoted).}

\item{value}{value to replace.}

\item{i, j}{indices specifying elements to extract or replace.}

\item{\dots}{Named or unnamed arguments to form a signature.}

\item{exact}{see \link[base]{Extract}}
}
\description{
The assessment and replacement methods for \link{APPconf-class}
}
\examples{
appconf$title
appconf[["title"]]
}
\concept{APPconf}
