% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableThemes.R
\name{availableThemes}
\alias{availableThemes}
\title{Available themes of color sets}
\usage{
availableThemes(output = "name")
}
\arguments{
\item{output}{The theme names or 'name' to print all available themes.}
}
\value{
A palette.
}
\description{
The color themes inspired by modern web design palettes
}
\examples{
availableThemes()

}
