% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataset.R
\name{createMetadata}
\alias{createMetadata}
\title{Create a metadata to describe the dataset
\enumerate{
\item must contain condition, cell type, tissue, species information
\item the conditions must be a keyword in database, used for whole database comparison
\item cell type must be standard words
}}
\usage{
createMetadata()
}
\description{
Create a metadata to describe the dataset
\enumerate{
\item must contain condition, cell type, tissue, species information
\item the conditions must be a keyword in database, used for whole database comparison
\item cell type must be standard words
}
}
