% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataset.R
\name{createSeuFromCellRanger}
\alias{createSeuFromCellRanger}
\title{load data from cellRanger}
\usage{
createSeuFromCellRanger(outsFolder)
}
\arguments{
\item{outsFolder}{the outs folder of cellRanger}
}
\description{
load data from cellRanger
}
