% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataset.R
\name{cteateSeuFromMatrix}
\alias{cteateSeuFromMatrix}
\title{load data from a count matrix}
\usage{
cteateSeuFromMatrix(matrix, meta, genes, cluster, ...)
}
\arguments{
\item{matrix}{count matrix}

\item{meta}{cell-level meta data}

\item{genes}{character. gene names, will be the rownames of the matrix}

\item{cluster}{the cluster coordinates}

\item{...}{The parameter passed to read.delim when read cluster file.}
}
\value{
An SeuratObject
}
\description{
load data from a count matrix
}
