% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyFiles.R
\name{extractFragmentNameMapList}
\alias{extractFragmentNameMapList}
\title{Extract fragment name map}
\usage{
extractFragmentNameMapList(obj, atacAssayName)
}
\arguments{
\item{obj}{input single-cell object for Seurat (v3+)}

\item{atacAssayName}{assay in single-cell data object to use for plotting
open chromatin.}
}
\value{
A list of fragment name map.
}
\description{
Try to extract cell name information. It will be used to map the fragment cell
name to the object cell name.
}
\examples{
library(Seurat)
#library(Signac)
#fnm <- extractFragmentNameMapList(atac_small, 'peaks')

}
