% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataset.R
\name{createDataSet}
\alias{createDataSet}
\title{Create a dataset
Create a dataset from an object}
\usage{
createDataSet(..., species, pmid, type, markers, keywords)
}
\description{
Create a dataset
Create a dataset from an object
}
