% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataset.R
\name{createDataSet}
\alias{createDataSet}
\title{Create a dataset
Create a dataset from a Seurat object}
\usage{
createDataSet(
  appconf,
  seu,
  config,
  contrast,
  assayName,
  gexSlot = c("data", "scale.data", "counts"),
  LOCKER = FALSE,
  datafolder = "data"
)
}
\arguments{
\item{appconf}{a list object represent the information about the dataset}

\item{seu}{a Seurat object}

\item{config}{config file for makeShinyFiles}

\item{contrast}{The contrast group}

\item{assayName}{assay in single-cell data object to use for plotting
gene expression, which must match one of the following:
\itemize{
\item{Seurat objects}: "RNA" or "integrated" assay,
default is "RNA"
}}

\item{gexSlot}{slot in single-cell assay to plot.
Default is to use the "data" slot}

\item{LOCKER}{Set locker if the file is required login}

\item{datafolder}{app data folder}
}
\value{
The updated Seurat object.
}
\description{
Create a dataset
Create a dataset from a Seurat object
}
\examples{
library(Seurat)
appconf <- createAppConfig(
           title="pbmc_small",
           destinationFolder = "pbmc_small",
           species = "Homo sapiens",
           doi="10.1038/nbt.3192",
           datatype = "scRNAseq")
createDataSet(appconf, pbmc_small, datafolder=tempdir())
}
