% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scInit.R
\name{scInit}
\alias{scInit}
\title{Create a scRNAseqApp project}
\usage{
scInit(
  app_path = getwd(),
  root = "admin",
  password = "scRNAseqApp",
  datafolder = "data",
  overwrite = FALSE,
  app_title = "scRNAseq Database",
  app_description =
    "This database is a collection of\\n        single cell RNA-seq data.",
  passphrase = NULL
)
}
\arguments{
\item{app_path}{path, a directory where do you want to create the app}

\item{root}{character(1), the user name for administrator}

\item{password}{character(1), the password for administrator}

\item{datafolder}{the folder where saved the dataset for the app}

\item{overwrite}{logical(1), overwrite the \code{app_path} if there is
a project.}

\item{app_title, app_description}{character(1). The title and
description of the home page.}

\item{passphrase}{A password to protect the data inside the database.}
}
\value{
no returns. This function will copy files to \code{app_path}
}
\description{
To run scRNAseqApp, you need to first create a directory
which contains the required files.
}
\examples{
if(interactive()){
    scInit()
}
}
