% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{scRNAseqApp}
\alias{scRNAseqApp}
\title{scRNAseqApp main function}
\usage{
scRNAseqApp(
  datafolder = "data",
  defaultDataset = "pbmc_small",
  windowTitle = "scRNAseq/scATACseq database",
  maxRequestSize = 1073741824,
  theme = bs_theme(bootswatch = "lumen"),
  use_bs_themer = FALSE,
  ...
)
}
\arguments{
\item{datafolder}{the folder where saved the dataset for the app}

\item{defaultDataset}{default dataset for the app.}

\item{windowTitle}{The title that should be displayed by the browser window.}

\item{maxRequestSize}{Maximal upload file size. Default is 1G.}

\item{theme}{A theme.}

\item{use_bs_themer}{logical(1). Used to determine the theme.}

\item{...}{parameters can be passed to shinyApp except ui and server.}
}
\value{
An object that represents the app.
}
\description{
create a scRNAseqApp once the initialization is done.
}
\examples{
if(interactive()){
  app_path=tempdir()
  scInit(app_path=app_path)
  setwd(app_path)
  scRNAseqApp()
}
}
