% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{scRNAseqApp}
\alias{scRNAseqApp}
\title{scRNAseqApp main function}
\usage{
scRNAseqApp(
  datafolder = "data",
  defaultDataset = "pbmc_small",
  windowTitle = "scRNAseq/scATACseq database",
  ...
)
}
\arguments{
\item{datafolder}{the folder where saved the dataset for the app}

\item{defaultDataset}{default dataset for the app.}

\item{windowTitle}{The title that should be displayed by the browser window.}

\item{...}{parameters can be passed to shinyApp except ui and server.}
}
\description{
scRNAseqApp main function
}
\examples{
if(interactive()){
  app_path=tempdir()
  scInit(app_path=app_path)
  setwd(app_path)
  scRNAseqApp()
}
}
